+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added setting to override nav link label.
 - Added setting to use chat window as a Utility.
 
+### Changed
+- Improved slideout which lists available skills.
+
 ### Fixed
 - Fixed bug which prevented the project config from updating properly.